Remove unnecessary cast

+++ src/storm/utility/bitoperations.h
@@ -3,7 +3,7 @@
 inline size_t smallestIntWithNBitsSet(size_t n) {
     static_assert(sizeof(size_t) == 8, "size_t has wrong size.");
     STORM_LOG_ASSERT(n < 64, "Input is too large."); // TODO fix this for 32 bit architectures!
-    if(n==0) return static_cast<size_t>(0);
+    if(n==0) return 0;
     return (1 << n) - 1;
 }
